iframe for flinkt cookie messaging; pure js toolbar
The toolbar is now in pure javascript.
The save button is not visible.

An invisible iframe now passes the current URI to flinkt as a hash.
This sadly reloads with each highlight since we remove and re-add the
toolbar.  Next step is to allow the snapshot code to "cut out" the
toolbar from the body innerHTML if possible, instead of removing it.